migrated to usePokemons for simpicity
